chore: refactor .dumi code folders (#53437)
