+++ .dumi/theme/common/SelectSemanticTemplate.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import useLocale from '../hooks/useLocale';
+import useLocale from '../../hooks/useLocale';
 import SemanticPreview from './SemanticPreview';
 
 export const locales = {
