+++ .dumi/theme/utils/tsToJs.ts
@@ -2,27 +2,16 @@ import * as ts from 'typescript';
 import { format } from '@prettier/sync';
 
 /**
- * TypeScript 到 JavaScript 代码转换工具
+ * 将 TypeScript 代码（含 TSX）转换为 JavaScript 代码（含 JSX）。
  *
- * 这个模块用于将 TypeScript 代码（包括 TSX）转换为 JavaScript 代码（包括 JSX）
- * 主要用于替代 sylvanas 库的功能，用于文档站点中的代码示例转换
+ * 使用 TypeScript 编译器 API 移除类型注解和类型导入，保留 JSX 语法和注释，并将代码转换为 JavaScript。转换结果会通过 Prettier 进行格式化，提升可读性。适用于文档示例、代码展示等场景。
  *
- * 实现原理：使用 TypeScript 编译器 API 将 TS 代码转换为 JS 代码
+ * @param tsCode - 输入的 TypeScript 代码字符串。
+ * @returns 转换并格式化后的 JavaScript 代码字符串。
  *
- * 特性：
- * 1. 删除所有类型注解
- * 2. 保留 JSX 语法
- * 3. 删除类型导入
- * 4. 转换 ES6+ 语法为更兼容的语法（如空值合并运算符）
- * 5. 保留原始代码风格和注释
- * 6. 使用 Prettier 格式化输出代码，提高可读性
- * 7. 处理 React 组件和 hooks 的转换
- * 8. 支持 TypeScript 特有语法（如装饰器、枚举等）的转换
- *
- * @param tsCode TypeScript 代码字符串
- * @returns 转换后的 JavaScript 代码
+ * @remark 若 Prettier 格式化失败，将返回未格式化的转换结果。
  */
-export default function (tsCode: string): string {
+export default function tsToJs(tsCode: string): string {
   // 设置编译器选项，保留 JSX 语法
   const compilerOptions: ts.CompilerOptions = {
     target: ts.ScriptTarget.ES2016, // 目标 ECMAScript 版本
@@ -60,16 +49,4 @@ export default function (tsCode: string): string {
     console.warn('Prettier 格式化出错:', error);
     return result.outputText;
   }
-}
-
-/**
- * 将 TypeScript 代码转换为 JavaScript 代码
- *
- * 这是一个公开的 API，供测试和外部调用使用
- *
- * @param tsCode TypeScript 代码字符串
- * @returns 转换后的 JavaScript 代码
- */
-export function parseText(tsCode: string): string {
-  return exports.default(tsCode);
 }
