+++ components/input/demo/_semantic_input.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { EditOutlined, UserOutlined } from '@ant-design/icons';
 import { Input } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/common/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
