+++ components/popconfirm/demo/_semantic.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Popconfirm } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/common/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
