+++ components/popover/demo/_semantic.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Popover } from 'antd';
 import type { PopoverProps } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/common/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
