+++ components/tree-select/demo/_semantic.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { TreeSelect } from 'antd';
 import type { TreeSelectProps } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/common/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
